+++ mex/editor/transform.py
@@ -41,7 +41,7 @@ def transform_value(value: object, allow_link: bool = True) -> EditorValue:
         )
     if isinstance(value, Identifier):
         return EditorValue(
-            text=str(value),
+            identifier=str(value),
             href=f"/item/{value}" if allow_link else None,
         )
     if isinstance(value, VocabularyEnum):