+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - create InputConfig for all fields, not just mex primary source
 - add `badge_default`, `badge_titles` and fine grained edit control to input config
 - allow editing for vocabulary fields and harmonize with link/text language editing
+- search and import ldap persons
+- resolve identifiers in search and edit view to their preferred title attributes
+- resolve identifiers in primary source search filter to their preferred title attribute
 
 ### Changes